requirements.txt
streamlit>=1.45.0
pandas>=2.2.0
openpyxl>=3.1.2
pdfplumber>=0.11.0
PyMuPDF>=1.24.0
Pillow>=10.3.0
opencv-python-headless>=4.10.0.84
pytesseract>=0.3.10
easyocr>=1.7.1
numpy>=1.26.4
fpdf2>=2.7.0
matplotlib>=3.8.0
